requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
dj-database-url==2.3.0
django-filter==25.2
psycopg2-binary==2.9.10
